+++ docs/meta/history.md
@@ -26,6 +26,10 @@ This is the first release candidate for Hatch v1, a complete rewrite.
 
 ### Unreleased
 
+***Added:***
+
+- Add option to clean build hook artifacts after each build
+
 ***Fixed:***
 
 - Properly include artifacts like C extensions that are built outside of package directories (with the intention of being placed directly inside `site-packages`) for projects with a `src`-layout structure